feat: 게시글에 이미지 등록 기능 추가

+++ src/main/java/kr/kakaotech/community/repository/PostImageRepository.java
@@ -3,5 +3,5 @@ package kr.kakaotech.community.repository;
 import kr.kakaotech.community.entity.PostImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostImageRepository extends JpaRepository<PostImage, Long> {
+public interface PostImageRepository extends JpaRepository<PostImage, Integer> {
 }
